created seperate routers for routes
